fix: fix types, deps types are implementation detail
